Moved window stuff into a package
