+++ src/ui/graphical/GraphilcalUI.java
@@ -1,5 +1,7 @@
 package ui.graphical;
 
+import ui.graphical.window.MainWindow;
+
 public class GraphilcalUI {
 	public static void becomeVisible() {
 		MainWindow main = new MainWindow();
